Move all classes to the rest package
