+++ modules/rest/src/main/java/org/apache/ignite/internal/rest/cluster/ClusterManagementRestFactory.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cluster.management.rest;
+package org.apache.ignite.internal.rest.cluster;
 
 import io.micronaut.context.annotation.Bean;
 import io.micronaut.context.annotation.Factory;
