+++ modules/rest/src/main/java/org/apache/ignite/internal/rest/cluster/TopologyController.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cluster.management.rest;
+package org.apache.ignite.internal.rest.cluster;
 
 import static java.util.stream.Collectors.toList;
