+++ modules/rest/src/main/java/org/apache/ignite/internal/rest/cluster/exception/handler/InvalidArgumentClusterInitializationExceptionHandler.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.cluster.management.rest.exception.handler;
+package org.apache.ignite.internal.rest.cluster.exception.handler;
 
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
@@ -23,24 +23,22 @@ import io.micronaut.http.HttpResponse;
 import io.micronaut.http.server.exceptions.ExceptionHandler;
 import jakarta.inject.Singleton;
 import org.apache.ignite.internal.rest.api.Problem;
+import org.apache.ignite.internal.rest.cluster.exception.InvalidArgumentClusterInitializationException;
 import org.apache.ignite.internal.rest.constants.HttpCode;
 import org.apache.ignite.internal.rest.problem.HttpProblemResponse;
-import org.apache.ignite.lang.IgniteInternalException;
 
 /**
- * Handles {@link IgniteInternalException} and represents it as a rest response.
+ * Handles {@link InvalidArgumentClusterInitializationException} and represents it as a rest response.
  */
 @Singleton
-@Requires(classes = {IgniteInternalException.class, ExceptionHandler.class})
-public class IgniteInternalExceptionHandler implements ExceptionHandler<IgniteInternalException, HttpResponse<? extends Problem>> {
+@Requires(classes = {InvalidArgumentClusterInitializationException.class, ExceptionHandler.class})
+public class InvalidArgumentClusterInitializationExceptionHandler implements
+        ExceptionHandler<InvalidArgumentClusterInitializationException, HttpResponse<? extends Problem>> {
 
     @Override
-    public HttpResponse<? extends Problem> handle(HttpRequest request, IgniteInternalException exception) {
+    public HttpResponse<? extends Problem> handle(HttpRequest request, InvalidArgumentClusterInitializationException exception) {
         return HttpProblemResponse.from(
-                Problem.fromHttpCode(HttpCode.INTERNAL_ERROR)
-                        .traceId(exception.traceId())
-                        .code(exception.codeAsString())
-                        .detail(exception.getMessage())
+                Problem.fromHttpCode(HttpCode.BAD_REQUEST).detail(exception.getCause().getMessage())
         );
     }
 }
