+++ modules/rest/src/main/java/org/apache/ignite/internal/rest/configuration/hocon/HoconPresentation.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.configuration.rest.presentation.hocon;
+package org.apache.ignite.internal.rest.configuration.hocon;
 
 import static com.typesafe.config.ConfigFactory.parseString;
 import static com.typesafe.config.ConfigRenderOptions.concise;
@@ -30,7 +30,7 @@ import org.apache.ignite.configuration.ConfigurationChangeException;
 import org.apache.ignite.configuration.validation.ConfigurationValidationException;
 import org.apache.ignite.internal.configuration.ConfigurationRegistry;
 import org.apache.ignite.internal.configuration.hocon.HoconConverter;
-import org.apache.ignite.internal.configuration.rest.presentation.ConfigurationPresentation;
+import org.apache.ignite.internal.rest.configuration.ConfigurationPresentation;
 import org.apache.ignite.lang.IgniteException;
 import org.jetbrains.annotations.Nullable;
