+++ modules/rest/src/main/java/org/apache/ignite/internal/rest/metrics/exception/handler/MetricNotFoundExceptionHandler.java
@@ -15,16 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.internal.metrics.rest.exception.handler;
+package org.apache.ignite.internal.rest.metrics.exception.handler;
 
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.server.exceptions.ExceptionHandler;
 import jakarta.inject.Singleton;
-import org.apache.ignite.internal.metrics.rest.exception.MetricNotFoundException;
 import org.apache.ignite.internal.rest.api.Problem;
 import org.apache.ignite.internal.rest.constants.HttpCode;
+import org.apache.ignite.internal.rest.metrics.exception.MetricNotFoundException;
 import org.apache.ignite.internal.rest.problem.HttpProblemResponse;
 
 /**
